pyproject.toml
[tool.poetry]
name = "dms-mowilex-automation"
version = "0.1.0"
description = "Playwright automation for DMS Mowilex Faktur Pajak download"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.40.0"
pytest = "^7.4.0"
pytest-playwright = "^0.4.0"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.5"
pdfminer-six = "^20260107"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--base-url https://dms.mowilex.id/ --browser chromium"
